refactor(tags-autocomplete): split filter-config helper into its own file and move out of pages-shared/shared
Mirrors the spanTypeFilter.tsx pattern: getTagsFilterConfig now lives at
v2/pages-shared/TagsAutocomplete/tagsFilterConfig.ts (separate from the
component). Also moves the directory from v2/pages-shared/shared/ up to
v2/pages-shared/ to match the convention used by sibling autocompletes
(DatasetVersionSelectBox, ExperimentFeedbackScoresViewer, etc.).
