+++ apps/opik-frontend/src/v2/pages-shared/dashboards/widgets/shared/ExperimentWidgetDataSection/ExperimentWidgetDataSection.tsx
@@ -24,7 +24,7 @@ import GroupsAccordionSection, {
 } from "@/shared/GroupsAccordionSection/GroupsAccordionSection";
 import DatasetSelectBox from "@/v2/pages-shared/experiments/DatasetSelectBox/DatasetSelectBox";
 import ExperimentsPathsAutocomplete from "@/v2/pages-shared/experiments/ExperimentsPathsAutocomplete/ExperimentsPathsAutocomplete";
-import { getTagsFilterConfig } from "@/v2/pages-shared/shared/TagsAutocomplete/TagsAutocomplete";
+import { getTagsFilterConfig } from "@/v2/pages-shared/TagsAutocomplete/tagsFilterConfig";
 import ExperimentFilterSelectBox from "./ExperimentFilterSelectBox";
 import { EXPERIMENT_IDS_FILTER_FIELD } from "@/lib/filters";
